+++ src/api/projects/create.rs
@@ -578,6 +578,9 @@ pub struct CreateProject<'a> {
     /// Set the access level for monitoring access.
     #[builder(default)]
     monitor_access_level: Option<FeatureAccessLevel>,
+    /// Set the access level for model experiment access.
+    #[builder(default)]
+    model_experiments_access_level: Option<FeatureAccessLevel>,
 
     /// Whether to enable email notifications or not.
     #[builder(default)]
@@ -931,6 +934,10 @@ impl<'a> Endpoint for CreateProject<'a> {
                 self.infrastructure_access_level,
             )
             .push_opt("monitor_access_level", self.monitor_access_level)
+            .push_opt(
+                "model_experiments_access_level",
+                self.model_experiments_access_level,
+            )
             .push_opt("emails_enabled", self.emails_enabled)
             .push_opt("show_default_award_emojis", self.show_default_award_emojis)
             .push_opt(
@@ -1718,6 +1725,28 @@ mod tests {
         let endpoint = CreateProject::builder()
             .name("name")
             .monitor_access_level(FeatureAccessLevel::Private)
+            .build()
+            .unwrap();
+        api::ignore(endpoint).query(&client).unwrap();
+    }
+
+    #[test]
+    fn endpoint_model_experiments_access_level() {
+        let endpoint = ExpectedUrl::builder()
+            .method(Method::POST)
+            .endpoint("projects")
+            .content_type("application/x-www-form-urlencoded")
+            .body_str(concat!(
+                "name=name",
+                "&model_experiments_access_level=private",
+            ))
+            .build()
+            .unwrap();
+        let client = SingleTestClient::new_raw(endpoint, "");
+
+        let endpoint = CreateProject::builder()
+            .name("name")
+            .model_experiments_access_level(FeatureAccessLevel::Private)
             .build()
             .unwrap();
         api::ignore(endpoint).query(&client).unwrap();